fixed some bugs occurring with large datafiles
git-svn-id: file:///Users/braun/svn/vermont/branches/vermont/filewriter@2122 aef3b71b-58ee-0310-9ba9-8811b9f0742f

+++ src/modules/ipfix/IpfixFileWriterCfg.hpp
@@ -32,7 +32,7 @@ private:
 	// config variables
 	std::string destinationPath;
 	std::string filenamePrefix;
-	uint64_t maximumFilesize;
+	uint32_t maximumFilesize;
 	uint16_t observationDomainId;
 };
